# Swagger_Documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Management API",
    "description": "APIs for managing employees",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/projects": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Get all projects",
        "description": "Retrieves a list of all projects in the system, including key details such as project name, deadline and associated department.",
        "operationId": "getProjects",
        "responses": {
          "200": {
            "description": "List of projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Create a new project",
        "description": "Allows users to create a new project by providing necessary details such as project name, deadline and associated department",
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Department not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}/assign": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "Assign employees to a project",
        "description": "Assigns a list of employees to a specific project, linking them to the project by their employee IDs.",
        "operationId": "assignEmployeesToProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project to which employees will be assigned",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "description": "List of employee IDs to assign to the project",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Employees successfully assigned to the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          },
          "404": {
            "description": "Project or one or more employees not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "One or more employees not in the same department as the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/employees": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Get employees assigned to a project",
        "description": "Retrieves a list of employees who are currently assigned to a specific project.",
        "operationId": "getAssignedEmployees",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project to retrieve assigned employees for",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          },
          "404": {
            "description": "Project not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectId}/employees/{employeeId}": {
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Remove an employee from a project",
        "description": "Removes a specific employee from a project using both the project ID and the employee ID.",
        "operationId": "removeEmployeeFromProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project from which to remove the employee",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "employeeId",
            "in": "path",
            "description": "ID of the employee to be removed from the project",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee successfully removed from the project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDTO"
                }
              }
            }
          },
          "404": {
            "description": "Project or employee not found for the given IDs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}": {
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Delete a project",
        "description": "Deletes a specific project by its ID, removing all associated data.",
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the project to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "404": {
            "description": "Project not found for the given ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected error occurred during input validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseDTO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "EmployeeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "joinDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "ProjectDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "departmentName": {
            "type": "string"
          }
        }
      },
      "CreateProjectDTO": {
        "type": "object",
        "description": "Details required to create a new project, including project name, description, etc.",
        "properties": {
          "name": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "deadline",
          "departmentId",
          "name"
        ]
      },
...
    }
  }
}